# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth_dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth_dto.LoginResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      refresh_token:
        type: string
      token:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  auth_dto.RefreshTokenResponse:
    properties:
      token:
        type: string
    type: object
  auth_dto.RegisterRequest:
    properties:
      address:
        type: string
      dob:
        type: string
      email:
        maxLength: 100
        type: string
      full_name:
        maxLength: 100
        type: string
      password:
        maxLength: 255
        minLength: 8
        type: string
      phone_number:
        maxLength: 15
        type: string
      username:
        maxLength: 20
        type: string
    required:
    - email
    - full_name
    - password
    - phone_number
    - username
    type: object
  auth_dto.RegisterResponse:
    properties:
      address:
        type: string
      dob:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      username:
        type: string
    type: object
  errors.Code:
    enum:
    - UNKNOWN_ERROR
    - INTERNAL_SERVER_ERROR
    - BAD_REQUEST
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - USER_NOT_FOUND
    - SESSION_NOT_FOUND
    - INVALID_PASSWORD
    - INVALID_STATUS_TRANSITION
    - TRANSACTION_NOT_FOUND
    - INSUFFICIENT_BALANCE
    type: string
    x-enum-varnames:
    - ErrCodeUnknownError
    - ErrCodeInternalServerError
    - ErrCodeBadRequest
    - ErrCodeUnauthorized
    - ErrCodeForbidden
    - ErrCodeNotFound
    - ErrCodeUserNotFound
    - ErrCodeSessionNotFound
    - ErrCodeInvalidPassword
    - ErrCodeInvalidStatusTransition
    - ErrCodeTransactionNotFound
    - ErrCodeInsufficientBalance
  response.BadRequestResponse:
    properties:
      error_code:
        $ref: '#/definitions/errors.Code'
      errors:
        items:
          $ref: '#/definitions/response.ValidationErrorField'
        type: array
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error_code:
        $ref: '#/definitions/errors.Code'
      message:
        type: string
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      meta: {}
    type: object
  response.ValidationErrorField:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Dokumentasi API Brilian.
  title: E Wallet API (User Management Service)
  version: "1.0"
paths:
  /user/login:
    post:
      consumes:
      - application/json
      description: Authenticates user credentials (username and password).
      parameters:
      - description: Payload login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth_dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login user
      tags:
      - Users
  /user/logout:
    delete:
      consumes:
      - application/json
      description: Logs out the authenticated user and invalidates the refresh token
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Logout user
      tags:
      - Users
  /user/refresh-token:
    put:
      consumes:
      - application/json
      description: Generates a new access token using a valid refresh token.
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth_dto.RefreshTokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Refresh token  user
      tags:
      - Users
  /user/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and securely stores the hashed password.
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth_dto.RegisterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register a new user
      tags:
      - Users
swagger: "2.0"